continue support for windows on gluten required libraries

+++ velox/common/base/BitUtil.h
@@ -94,7 +94,7 @@ constexpr inline T roundUp(T value, U factor) {
 }
 
 constexpr inline uint64_t lowMask(int32_t bits) {
-  return (1UL << bits) - 1;
+  return (uint64_t(1) << bits) - 1;
 }
 constexpr inline uint64_t lowMask(size_t bits) {
   return (size_t(1) << bits) - 1;
